AutoloadGenerator: Most methods that depend on filesystem stuff are factored out into AutoloadGeneratorHelper.
